move classes of view task log to right directory
