+++ src/test/kotlin/logic/usecases/task/ViewTaskLogsUseCaseTest.kt
@@ -1,10 +1,8 @@
-package usecase
+package logic.usecases.task
 
 import io.mockk.every
 import io.mockk.mockk
-import logic.exception.PlanMateException.NotFoundException.TaskNotFoundException
-
-import logic.exception.PlanMateException.ValidationException.InvalidTaskIDException
+import logic.exception.PlanMateException
 import logic.repository.LogRepository
 import logic.usecases.ViewTaskLogsUseCase
 import org.junit.jupiter.api.BeforeEach
@@ -12,9 +10,8 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import utils.hepler.taskLogs
-import utils.hepler.validId
-import java.util.UUID
+import logic.usecases.testFactory.taskLogs
+import logic.usecases.testFactory.validId
 import kotlin.test.assertEquals
 
 class ViewTaskLogsUseCaseTest {
@@ -48,16 +45,16 @@ class ViewTaskLogsUseCaseTest {
     )
     fun `should throw InvalidTaskIDException when task id is invalid UUID format`(id:String){
 
-        assertThrows<InvalidTaskIDException>{
+        assertThrows<PlanMateException.ValidationException.InvalidTaskIDException> {
             useCase.viewTaskLogs(id)
         }
     }
 
     @Test
     fun `should throw TaskNotFoundException when id of task not found`(){
-        every { repository.viewLogsById(validId) } throws TaskNotFoundException
+        every { repository.viewLogsById(validId) } throws PlanMateException.NotFoundException.TaskNotFoundException
 
-        assertThrows<TaskNotFoundException> {
+        assertThrows<PlanMateException.NotFoundException.TaskNotFoundException> {
             useCase.viewTaskLogs(validId.toString())
 
         }
